preprocessing/requirements.txt
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
Automat==24.8.1
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.38.35
botocore==1.38.35
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.10
dagshub-annotation-converter==0.1.9
databricks-sdk==0.56.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-play-scraper==1.2.7
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lxml==5.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mlflow==3.1.0
mlflow-skinny==3.1.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.29.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.3
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pathvalidate==3.2.3
pillow==11.1.0
pipenv==2025.0.3
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.2
Protego==0.4.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.23.1
rsa==4.9.1
s3transfer==0.13.0
Sastrawi==1.0.1
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.2
Scrapy==2.12.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.43.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelib==1.7.1
Twisted==24.11.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.3
virtualenv==20.29.3
w3lib==2.3.1
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
yarl==1.20.1
yellowbrick==1.5
zipp==3.23.0
zope.interface==7.2
